src/requirements.txt
Django==1.11
Pillow
django-contrib-comments
django-debug-toolbar
easy-thumbnails
django-taggit
gunicorn==19.7.1
psycopg2==2.7
djangorestframework
